Improve validation script
Improvements include:

* Accept template(s) via arguments rather than hardcoded
* Can send a single resource to OPA or all from directory
* Add --test flag to run integration tests (same as before)
* Without --test flag, just print the response from OPA, allowing
  rapid iteration of policy and template development.
* Accept JSON templates as well as YAML.
* Fixed some strict failures due to unused imports,
  not sure how they passed the CI tests before

Fixed #20 and #21
