+++ policy/eks/cluster/logging_enabled_test.rego
@@ -1,15 +1,13 @@
 package aws.eks.logging_enabled_test
 
-import data.aws.eks.cluster.deny
-
 import future.keywords
 
-import data.assertions.assert_empty
+import data.aws.eks.cluster.deny
+
 import data.assertions.assert_in
 import data.assertions.assert_not_in
 
 import data.test_helpers.create_with_properties
-import data.test_helpers.with_properties
 
 test_allow_cluster_logging_enabled {
 	inp := create_with_properties("AWS::EKS::Cluster", "EksCluster", {
